+++ pkg/ctl/upgrade/cluster.go
@@ -1,22 +1,16 @@
 package upgrade
 
 import (
-	"fmt"
 	"time"
 
-	"github.com/weaveworks/eksctl/pkg/eks"
-
-	"github.com/weaveworks/eksctl/pkg/cfn/manager"
+	"github.com/weaveworks/eksctl/pkg/cluster"
 
 	"github.com/kris-nova/logger"
-	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 
 	api "github.com/weaveworks/eksctl/pkg/apis/eksctl.io/v1alpha5"
 	"github.com/weaveworks/eksctl/pkg/ctl/cmdutils"
-	"github.com/weaveworks/eksctl/pkg/printers"
-	"github.com/weaveworks/eksctl/pkg/utils"
 )
 
 // updating from 1.15 to 1.16 has been observed to take longer than the default value of 25 minutes
@@ -85,158 +79,10 @@ func DoUpgradeCluster(cmd *cmdutils.Cmd) error {
 		logger.Warning("NOTE: cluster VPC (subnets, routing & NAT Gateway) configuration changes are not yet implemented")
 	}
 
-	currentVersion := ctl.ControlPlaneVersion()
-	versionUpdateRequired, err := requiresVersionUpgrade(cfg.Metadata, currentVersion)
+	c, err := cluster.New(cfg, ctl)
 	if err != nil {
 		return err
 	}
 
-	stackManager := ctl.NewStackManager(cfg)
-	stacks, err := stackManager.DescribeStacks()
-	if err != nil {
-		return err
-	}
-
-	if manager.IsClusterStack(stacks) {
-		return upgradeEKSCTLCluster(versionUpdateRequired, cmd.Plan, cfg, currentVersion, ctl, stackManager)
-	}
-
-	return upgradeNonEKSCTLCluster(versionUpdateRequired, cmd.Plan, cfg, currentVersion, ctl)
-}
-
-func upgradeEKSCTLCluster(versionUpdateRequired, dryRun bool, cfg *api.ClusterConfig, currentVersion string, ctl *eks.ClusterProvider, stackManager *manager.StackCollection) error {
-	if err := ctl.LoadClusterVPC(cfg); err != nil {
-		return errors.Wrapf(err, "getting VPC configuration for cluster %q", cfg.Metadata.Name)
-	}
-
-	printer := printers.NewJSONPrinter()
-	if err := printer.LogObj(logger.Debug, "cfg.json = \\\n%s\n", cfg); err != nil {
-		return err
-	}
-
-	if versionUpdateRequired {
-		if err := updateVersion(dryRun, cfg, currentVersion, ctl); err != nil {
-			return err
-		}
-	}
-
-	if err := ctl.RefreshClusterStatus(cfg); err != nil {
-		return err
-	}
-
-	supportsManagedNodes, err := ctl.SupportsManagedNodes(cfg)
-	if err != nil {
-		return err
-	}
-
-	stackUpdateRequired, err := stackManager.AppendNewClusterStackResource(dryRun, supportsManagedNodes)
-	if err != nil {
-		return err
-	}
-
-	if err := ctl.ValidateExistingNodeGroupsForCompatibility(cfg, stackManager); err != nil {
-		logger.Critical("failed checking nodegroups", err.Error())
-	}
-
-	cmdutils.LogPlanModeWarning(dryRun && (stackUpdateRequired || versionUpdateRequired))
-	return nil
-}
-
-func upgradeNonEKSCTLCluster(versionUpdateRequired, dryRun bool, cfg *api.ClusterConfig, currentVersion string, ctl *eks.ClusterProvider) error {
-	printer := printers.NewJSONPrinter()
-	if err := printer.LogObj(logger.Debug, "cfg.json = \\\n%s\n", cfg); err != nil {
-		return err
-	}
-
-	if versionUpdateRequired {
-		if err := updateVersion(dryRun, cfg, currentVersion, ctl); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func updateVersion(dryRun bool, cfg *api.ClusterConfig, currentVersion string, ctl *eks.ClusterProvider) error {
-	msgNodeGroupsAndAddons := "you will need to follow the upgrade procedure for all of nodegroups and add-ons"
-	cmdutils.LogIntendedAction(dryRun, "upgrade cluster %q control plane from current version %q to %q", cfg.Metadata.Name, currentVersion, cfg.Metadata.Version)
-	if !dryRun {
-		if err := ctl.UpdateClusterVersionBlocking(cfg); err != nil {
-			return err
-		}
-		logger.Success("cluster %q control plane has been upgraded to version %q", cfg.Metadata.Name, cfg.Metadata.Version)
-		logger.Info(msgNodeGroupsAndAddons)
-	}
-	return nil
-}
-
-func requiresVersionUpgrade(clusterMeta *api.ClusterMeta, currentEKSVersion string) (bool, error) {
-	nextVersion, err := getNextVersion(currentEKSVersion)
-	if err != nil {
-		return false, err
-	}
-
-	// If the version was not specified default to the next Kubernetes version and assume the user intended to upgrade if possible
-	// also support "auto" as version (see #2461)
-	if clusterMeta.Version == "" || clusterMeta.Version == "auto" {
-		if api.IsSupportedVersion(nextVersion) {
-			clusterMeta.Version = nextVersion
-			return true, nil
-		}
-
-		// There is no new version, stay in the current one
-		clusterMeta.Version = currentEKSVersion
-		return false, nil
-	}
-
-	if c, err := utils.CompareVersions(clusterMeta.Version, currentEKSVersion); err != nil {
-		return false, errors.Wrap(err, "couldn't compare versions for upgrade")
-	} else if c < 0 {
-		return false, fmt.Errorf("cannot upgrade to a lower version. Found given target version %q, current cluster version %q", clusterMeta.Version, currentEKSVersion)
-	}
-
-	if api.IsDeprecatedVersion(clusterMeta.Version) {
-		return false, fmt.Errorf("control plane version %q has been deprecated", clusterMeta.Version)
-	}
-
-	if !api.IsSupportedVersion(clusterMeta.Version) {
-		return false, fmt.Errorf("control plane version %q is not known to this version of eksctl, try to upgrade eksctl first", clusterMeta.Version)
-	}
-
-	if clusterMeta.Version == currentEKSVersion {
-		return false, nil
-	}
-
-	if clusterMeta.Version == nextVersion {
-		return true, nil
-	}
-
-	return false, fmt.Errorf(
-		"upgrading more than one version at a time is not supported. Found upgrade from %q to %q. Please upgrade to %q first",
-		currentEKSVersion,
-		clusterMeta.Version,
-		nextVersion)
-}
-
-func getNextVersion(currentVersion string) (string, error) {
-	switch currentVersion {
-	case "":
-		return "", errors.New("unable to get control plane version")
-	case api.Version1_12:
-		return api.Version1_13, nil
-	case api.Version1_13:
-		return api.Version1_14, nil
-	case api.Version1_14:
-		return api.Version1_15, nil
-	case api.Version1_15:
-		return api.Version1_16, nil
-	case api.Version1_16:
-		return api.Version1_17, nil
-	case api.Version1_17:
-		return api.Version1_18, nil
-	case api.Version1_18:
-		return api.Version1_19, nil
-	default:
-		// version of control plane is not known to us, maybe we are just too old...
-		return "", fmt.Errorf("control plane version %q is not known to this version of eksctl, try to upgrade eksctl first", currentVersion)
-	}
+	return c.Upgrade(cmd.Plan)
 }
